refactor: spring time cleaning
After a hole year of using Kickstarter as a startingpoint, it was time for spring cleaning.

This is a total rewrite of my current setup. All (core) configurations is now onder `lua/*.lua`. Plugins can be found
under `lua/plugins` and extras (e.g. language specific or formatting) can be found under `lua/extras`.

The extras can be enabled/disabled with `Extras` command.